Improved translations for CollectorNumber in Italian and Spanish, and while at it...
...I got rid of duplication for the lblSet col, while also improving translations.
In particular, the "lblSetEdition" column was translated as the "Mistery column" (for some reason :D), whereas a second lblSet2 was instead defined and used, namely "the label tooltip for the Set Columndef".
On this note, this col was also mistranslated in ES, IT, DE - those are the language I could handle so far.
In fact "Set" was intended (and so translated) as "to set" instead as for "Expansion/Edition".
Corrected accordingly! :)