Unban War Elephant
For simplicity I will call AI:RemoveDeck:Random as the same thing as "Banned" for the AI.

War Elephant is banned from AI draft decks because it has banding.

The AI does not know how to use banding intelligently, but has no trouble casting, attacking, and blocking with War Elephant, even if it's not optimal. Giving white a 2/2 Trample for 4 is still better than nothing.

Additionally, right now 5 of 11 white cards are banned in ARN. This makes the AI that plays white decks short on playables consistently.

Therefore, this unbans War Elephant given that it's better to give the AI a creature it can play sub-optimally, rather than give it nothing at all.