Extended DB API with pair of new methods
GetCardFromEditions (series of) methods now support to specify
a Predicate to filter the set of available candidates when gathering all the same card prints (by name).

For this reason, a new Predicate (isLegal) has been added to include limitations (filtering) based on set codes.

getAllCards method now include the possibility to provide a predicate
as a filter.