DeckRecognizer.Token to include whether request in card list had any set code
This commit adds in a new piece of information to Token in DeckRecognizer to mark whether original card request included or not a set code.
This flag will be only used by actual CardToken and will be leveraged later on by "smart card optimisation option" to be added to DeckImport on desktop too.

Majority of tests in DeckRecognizerTest have been updated to cross-check that no bug has been unintentionally introduced.