Import/Replace deck after import in Deck Editor
The editor has been slightly change to support the new modality in deck Importer to replace/import cards into current deck (if any).
Also, this commit fixes a bug also considering deck name to be replaced if deck from importer has a name.

To do so, the DeckImport won't take the callback to run directly into the constructor, but the callback will be fed into with a separated method. This is because the callback would need to access the dialog instance to access which option was selected by the user about replace/import.