Fix simulation AI logic related to playing lands.
This got broken by the following refactor:
https://github.com/Card-Forge/forge/commit/79c9c914e21521c03ed219954810105fac3a639a

The result was that simulation AI was ignoring certain decision trees that involved
playing lands, leading to not considering certain lines of play and some log
messages printed to standard error. I've added a test that covers this logic to prevent
it breaking again in the future.

Also, a couple small clean ups to related tests.