Update leonardo_da_vinci.txt
I think this should be correct, based on my reading of other scripts that use both Remember and Imprint, like Break Out, but I'm not super well versed on the differences and relations between those two things. I have tested it both with and without Rest in Peace in play and it does seem to work as it should, though.