Add files via upload (#5486)
Attempting to add UNF's Nearby Planet using the Omo effect from M3C (and to a lesser extent Planar Nexus from MH3) as a reference.

Tested card, seems to work as intended. The typeline is messier than with the Mistform Ultimus/changeling effect for creatures (Mistform displays "(All)" whereas the Omo version spells out the types), but it is consistent with the referenced Modern Horizons cards.